Fixed support for git branch that was lost in refactoring
